requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
rembg[new]==2.0.50
numpy==1.26.2
gradio==4.7.1
python-jose[cryptography]==3.3.0
requests==2.31.0
python-dotenv==1.0.0 